Extract reifiers/has_one
This is the first file extracted from reifier.rb. I anticipate
extracting files for each type of association: has one, has many, etc.

Having a separate file allows us to organize related code. For example,
the `without_persisting` method is only used in HasOne.
